requirements.txt
newspaper3k
TextBlob
Scrapy-py3
